[level 3] Title: 연속 펄스 부분 수열의 합, Time: 50.41 ms, Memory: 140 MB -BaekjoonHub

+++ 프로그래머스/3/161988. 연속 펄스 부분 수열의 합/연속 펄스 부분 수열의 합.java
@@ -2,13 +2,13 @@ import java.util.*;
 
 class Solution {
     public long solution(int[] sequence) {
-        List<Integer> v = new ArrayList<>();
+        List<Long> v = new ArrayList<>();
         long answer = 0, mi = 0, ma = 0;
         for (int i = 0; i < sequence.length; i++) {
             if (i % 2 == 1) {
                 sequence[i] *= -1;
             }
-            v.add(sequence[i]);
+            v.add((long) sequence[i]);
             if (i > 0) {
                 v.set(i, v.get(i) + v.get(i - 1));
             }
